verifying update on pdv using config local and network server, retaguarda using DB and config, download package only in server

+++ src/main/java/com/g2soft/g2portal/service/dropbox/DropboxDeleteFile.java
@@ -1,7 +1,9 @@
 package com.g2soft.g2portal.service.dropbox;
 
+import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -24,19 +26,23 @@ private static final String DROPBOX_ACCESS_TOKEN = "qG4Cd10pSGAAAAAAAAAADmqNd19b
 	DbxRequestConfig configDropbox = new DbxRequestConfig("dropbox/g2adv");
 	DbxClientV2 clientDropbox = new DbxClientV2(configDropbox, DROPBOX_ACCESS_TOKEN);
 
+	private static final String LOG_PATH = "C:\\G2 Soft\\logs\\logG2Portal.txt";
+	
 	private static final Logger logger = (Logger) LogManager.getLogger(DropboxDeleteFile.class.getName());
 	
 	public void deleteFile(String path, JLabel labelTaskStatus) {
 		try {
-			clientDropbox.files().getMetadata(path);
-			labelTaskStatus.setText("Deletando xml cancelada: " + path);
-			clientDropbox.files().deleteV2(path);
-			writeDeleteOnLog("arquivo_deletado=" + path);
+			if (canDeleteFile(path)) {
+				clientDropbox.files().getMetadata(path);
+				labelTaskStatus.setText("Deletando arquivo: " + path);
+				clientDropbox.files().deleteV2(path);
+				writeDeleteOnLog("arquivo_deletado=" + path);
+			}			
 		} catch (GetMetadataErrorException e) {
-			logger.error(e);
-			e.printStackTrace();
 			if (e.errorValue.isPath() && e.errorValue.getPathValue().isNotFound()) {
-				System.out.println("Xml não encontrada: " + path);
+				System.out.println("Arquivo não encontrado: " + path);
+				logger.error("Arquivo não encontrado: " + path);
+				writeDeleteOnLog("arquivo_nao_encontrado=" + path);
 			}
 		} catch (DbxException e) {
 			logger.error(e);
@@ -44,9 +50,36 @@ private static final String DROPBOX_ACCESS_TOKEN = "qG4Cd10pSGAAAAAAAAAADmqNd19b
 		}
 	}
 	
+	private boolean canDeleteFile(String path) {
+		BufferedReader buffReader = null;
+		try {
+			File fileLog = new File(LOG_PATH);
+			buffReader = new BufferedReader(new FileReader(fileLog));
+			String currentLine;
+			while((currentLine = buffReader.readLine()) != null) {
+				if ((currentLine.contains("arquivo_nao_encontrado") || currentLine.contains("arquivo_deletado")) &&
+						currentLine.contains(path)) {
+					return false;
+				}
+			}
+			buffReader.close();
+		} catch (IOException e) {
+			e.printStackTrace();
+		} finally {
+			if (buffReader != null)
+				try {
+					buffReader.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+		}
+		
+		return true;
+	}
+	
 	public void writeDeleteOnLog(String text) {
 		try {
-			File logFile = new File("C:\\G2 Soft\\logs\\logG2Portal.txt");			
+			File logFile = new File(LOG_PATH);			
 			SimpleDateFormat dataFormat = new SimpleDateFormat("dd/MM/yyyy HH:mm:ss");
 			String strDateNow = dataFormat.format(new Date());
 			if (!logFile.exists()) 
@@ -55,7 +88,7 @@ private static final String DROPBOX_ACCESS_TOKEN = "qG4Cd10pSGAAAAAAAAAADmqNd19b
 			output.newLine();
 			output.write(text + ", data=" + strDateNow);
 			output.flush();
-			output.close();			
+			output.close();	
 		} catch (IOException e) {
 			logger.error(e);
 		}
